Add route test for hearing-location page
