+++ src/main/features/directions-questionnaire/views/hearing-location.njk
@@ -12,7 +12,7 @@
         <fieldset id="otherWitnesses[label]" class="form-group {% if form.hasErrors() %} form-group-error {% endif %}">
           <legend>
             <span class="visuallyhidden">
-              {{ t('Do you want witnesses to give evidence?') }}
+              {{ t('Choose a hearing location') }}
             </span>
           </legend>
